ACTIN-3233: Move Logging utils from common to actin-utils
