+++ common/pom.xml
@@ -23,6 +23,10 @@
         </dependency>
 
         <dependency>
+            <groupId>com.hartwig.actin</groupId>
+            <artifactId>actin-utils</artifactId>
+        </dependency>
+        <dependency>
             <groupId>org.jetbrains.kotlin</groupId>
             <artifactId>kotlin-stdlib</artifactId>
         </dependency>
