+++ database/src/main/kotlin/com/hartwig/actin/database/algo/TreatmentMatchLoaderConfig.kt
@@ -4,7 +4,7 @@ import com.hartwig.actin.database.DatabaseLoaderConfig
 import com.hartwig.actin.util.ApplicationConfig
 import org.apache.commons.cli.CommandLine
 import org.apache.commons.cli.Options
-import com.hartwig.actin.util.enableDebugLogging
+import com.hartwig.actin.utils.enableDebugLogging
 import io.github.oshai.kotlinlogging.KotlinLogging
 
 data class TreatmentMatchLoaderConfig(
