+++ molecular/src/main/kotlin/com/hartwig/actin/molecular/cancerassociatedvariantcomparison/CancerAssociatedVariantComparisonConfig.kt
@@ -3,7 +3,7 @@ package com.hartwig.actin.molecular.cancerassociatedvariantcomparison
 import com.hartwig.actin.util.ApplicationConfig
 import org.apache.commons.cli.CommandLine
 import org.apache.commons.cli.Options
-import com.hartwig.actin.util.enableDebugLogging
+import com.hartwig.actin.utils.enableDebugLogging
 import io.github.oshai.kotlinlogging.KotlinLogging
 
 data class CancerAssociatedVariantComparisonConfig(
