+++ pom.xml
@@ -28,7 +28,7 @@
 
         <actin-datamodel.version>3.0.0</actin-datamodel.version>
         <actin-treatment-database.version>0.4.2</actin-treatment-database.version>
-        <actin-utils.version>2.0.0</actin-utils.version>
+        <actin-utils.version>2.1.0-beta.1</actin-utils.version>
         <serve.version>8.6.0</serve.version>
         <orange-datamodel.version>3.0.0</orange-datamodel.version>
         <actin-transvar.version>2.0.0</actin-transvar.version>
