+++ report/src/main/kotlin/com/hartwig/actin/report/ReporterConfig.kt
@@ -6,7 +6,7 @@ import com.hartwig.actin.util.ApplicationConfig
 import java.time.LocalDate
 import org.apache.commons.cli.CommandLine
 import org.apache.commons.cli.Options
-import com.hartwig.actin.util.enableDebugLogging
+import com.hartwig.actin.utils.enableDebugLogging
 import io.github.oshai.kotlinlogging.KotlinLogging
 
 data class ReporterConfig(
